requirements.txt
attrs==22.2.0
bcrypt==4.0.1
black==23.1.0
cfgv==3.3.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
exceptiongroup==1.1.0
Faker==16.7.0
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
identify==2.5.17
importlib-metadata==6.0.0
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==3.0.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytest==7.2.1
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
tomli==2.0.1
typing-extensions==4.4.0
virtualenv==20.18.0
Werkzeug==2.2.2
zipp==3.12.1
